backend/requirements.txt
# Flask and Extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Session>=0.5.0   # Updated for Flask 2.3 compatibility

# WebSocket Support (Real-time updates)
Flask-SocketIO==5.5.1  # Compatible with Flask 2.3.3
python-socketio==5.11.0  # Compatible with Flask-SocketIO 5.5.1
python-engineio==4.9.0  # Compatible with python-socketio 5.11.0
simple-websocket==1.0.0  # Required for WebSocket transport

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.23
alembic==1.12.1

# Environment and Configuration
python-dotenv==1.0.0

# Web Scraping and HTTP Requests
requests==2.31.0
beautifulsoup4==4.12.2

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# Development Dependencies (optional)
Flask-Testing==0.8.1
pytest==7.4.3
pytest-flask==1.2.0

# Production Dependencies (optional)
gunicorn==21.2.0
redis==5.0.1

# OAuth
authlib==1.2.0

# PDF Generation
pdfkit==1.0.0
weasyprint==60.1

# Face Recognition (OpenCV)
opencv-contrib-python==4.8.1.78
Pillow==10.2.0
mailersend==0.5.7

reportlab==4.0.7 
num2words==0.5.13
weasyprint==60.1